requirements.txt
torch==1.13.1
pandas==1.5.2
scikit-learn==1.2.0
numpy==1.24.1
timm==0.6.12
hydra-core==1.3.1
wandb==0.13.10
torchmetrics==0.11.1